+++ object.cc
@@ -23,7 +23,7 @@ void object::set_conduc(double c) {
 
 medium::medium() {};
 
-bool medium::inside(vector<double> p) {
+bool medium::inside(vector<int> p) {
     return true;
 }
 
@@ -31,7 +31,7 @@ bool medium::inside(vector<double> p) {
 
 cylinder::cylinder(double x, double y, double r): x(x), y(y), r(r) {};
 
-bool cylinder::inside(vector<double> p){
+bool cylinder::inside(vector<int> p){
     if (pow(p[0]-x,2) + pow(p[1]-y,2) < pow(r,2))
        return true;
     else return false; 
@@ -40,7 +40,7 @@ bool cylinder::inside(vector<double> p){
 
 rectangle::rectangle(vector<double> pa, vector<double> pb): pa(pa), pb(pb) {};
 
-bool rectangle::inside(vector<double> p) {
+bool rectangle::inside(vector<int> p) {
     if (p[0] < pa[0] or p[0] > pb[0] or p[1] < pa[1] or p[1] > pb[1]) 
         return false;
     else return true;
