+++ source.cc
@@ -25,27 +25,27 @@ void source::set_F(Field2D *field) {
     t = &(F->t);
 }
 
-continuous_point_source::continuous_point_source(vector<double> p, double freq): p(p), freq(freq) {};
+continuous_point_source::continuous_point_source(vector<int> p, double freq): p(p), freq(freq) {};
 void continuous_point_source::pulse() {
     static vector<int> pi = (F->grid).convertToGrid(p);
     F->Ez[pi[0]][pi[1]] += sin(2*M_PI*freq*(*t));
 }
 
-gaussian_point_source::gaussian_point_source(vector<double> p, double T0, double sig):
+gaussian_point_source::gaussian_point_source(vector<int> p, double T0, double sig):
     p(p), T0(T0), sig(sig) {};
 void gaussian_point_source::pulse() {
     static vector<int> pi = (F->grid).convertToGrid(p);
     F->Ez[pi[0]][pi[1]] += exp(-0.5*(pow((*t-T0)/sig,2)));
 }
 
-custom_point_source::custom_point_source(vector<double> p, double (*time_func)(double)):
+custom_point_source::custom_point_source(vector<int> p, double (*time_func)(double)):
     p(p), time_func(time_func) {};
 void custom_point_source::pulse() {
     static vector<int> pi = (F->grid).convertToGrid(p);
     F->Ez[pi[0]][pi[1]] += time_func(*t);
 }
     
-continuous_line_source::continuous_line_source(vector<double> p1, vector<double> p2, double freq): p1(p1), p2(p2), freq(freq) {};
+continuous_line_source::continuous_line_source(vector<int> p1, vector<int> p2, double freq): p1(p1), p2(p2), freq(freq) {};
 void continuous_line_source::pulse() {
     static vector<int> p1i = (F->grid).convertToGrid(p1);
     static vector<int> p2i = (F->grid).convertToGrid(p2);
@@ -70,12 +70,12 @@ void continuous_line_source::pulse() {
     }
 }
 
-gaussian_line_source::gaussian_line_source(vector<double> p1, vector<double> p2, double T0, double sig):
+gaussian_line_source::gaussian_line_source(vector<int> p1, vector<int> p2, double T0, double sig):
     p1(p1), p2(p2),T0(T0), sig(sig) {};
 void gaussian_line_source::pulse() {
 }
 
-custom_line_source::custom_line_source(vector<double> p1, vector<double> p2, double (*time_func)(double)):
+custom_line_source::custom_line_source(vector<int> p1, vector<int> p2, double (*time_func)(double)):
     p1(p1), p2(p2), time_func(time_func) {};
 void custom_line_source::pulse() {
 }
