+++ source.h
@@ -28,56 +28,56 @@ public:
 
 class continuous_point_source: public source {
 public:
-    std::vector<double> p;
+    std::vector<int> p;
     double freq;
 public:
-    continuous_point_source(std::vector<double> p, double f);
+    continuous_point_source(std::vector<int> p, double f);
     void pulse();
 };
 
 class gaussian_point_source: public source {
 public:
-    std::vector<double> p;
+    std::vector<int> p;
     double T0, sig;
 public:
-    gaussian_point_source(std::vector<double> p, double T0, double sig);
+    gaussian_point_source(std::vector<int> p, double T0, double sig);
     void pulse();
 };
 
 class custom_point_source: public source {
 public:
-    std::vector<double> p;
+    std::vector<int> p;
     double (*time_func)(double);
 public:
-    custom_point_source(std::vector<double> p, double (*time_func)(double));
+    custom_point_source(std::vector<int> p, double (*time_func)(double));
     void pulse();
 };
 
 
 class continuous_line_source: public source {
 public:
-    std::vector<double> p1, p2;
+    std::vector<int> p1, p2;
     double freq;
 public:
-    continuous_line_source(std::vector<double> p1, std::vector<double> p2, double f);
+    continuous_line_source(std::vector<int> p1, std::vector<int> p2, double f);
     void pulse();
 };
 
 class gaussian_line_source: public source {
 public:
-    std::vector<double> p1, p2;
+    std::vector<int> p1, p2;
     double T0, sig;
 public:
-    gaussian_line_source(std::vector<double> p1, std::vector<double> p2, double T0, double sig);
+    gaussian_line_source(std::vector<int> p1, std::vector<int> p2, double T0, double sig);
     void pulse();
 };
 
 class custom_line_source: public source {
 public:
-    std::vector<double> p1, p2;
+    std::vector<int> p1, p2;
     double (*time_func)(double);
 public:
-    custom_line_source(std::vector<double> p1, std::vector<double> p2, double (*time_func)(double));
+    custom_line_source(std::vector<int> p1, std::vector<int> p2, double (*time_func)(double));
     void pulse();
 };
